Make UI and chat UI more streamlined
